Added bk shell function, updated ssh sock
SSH auth socket variable was set regardless of whether it was already in
use, this changes that so that it only sets that variable if it isn't
already set.

Also added the bk function, which automatically adds > /dev/null 2>&1 &
to the end of the entered command to make for forking GUI programs in a
current directory cleanly easier.

Also removed the synaptic alias, as I haven't used Ubuntu in ages, and
don't intend to go back soon.